experiment_tools/requirements.txt
absl-py==2.1.0
apricot-select==0.6.1
asciitree==0.3.3
astunparse==1.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
cyvcf2==0.30.28
dask==2024.3.0
dm-tree==0.1.8
fasteners==0.19
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
hmmlearn==0.3.2
humanfriendly==10.0
idna==3.6
importlib_metadata==7.0.2
Jinja2==3.1.3
joblib==1.3.2
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
llvmlite==0.42.0
locket==1.0.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
nose==1.3.7
numba==0.59.0
numcodecs==0.12.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
partd==1.4.1
pillow==10.2.0
pomegranate==1.0.4
POT==0.9.3
protobuf==4.25.3
Pygments==2.17.2
pyparsing==3.1.2
pysam==0.22.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
resample==1.7.1
rich==13.7.1
scikit-allel==1.3.7
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
toolz==0.12.1
torch==2.2.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vcfpy==0.13.8
Werkzeug==3.0.1
wrapt==1.16.0
zarr==2.17.1
zipp==3.18.0
